+++ src/seq2seq.py
@@ -183,7 +183,7 @@ class Decoder:
             char_id = score.argmax(axis=2)
             sampled.append(char_id.flatten())
 
-        return np.array(sampled).T
+        return np.array(sampled, dtype=np.int).T
 
     def generate_with_cf(self, h, start_id, sample_size):
         '''
@@ -238,7 +238,7 @@ class Decoder:
 
         cf = sum_cf / counts  # mean
 
-        return np.array(sampled).T, cf
+        return np.array(sampled, dtype=np.int).T, cf
 
 
 class Seq2seq(BaseModel):
